stripe-integration/requirements.txt
flask==3.1.0
stripe==8.11.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.4
pytest-cov==6.0.0
